backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
alembic==1.17.2
    # via -r requirements.in
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   starlette
    #   watchfiles
asyncpg==0.31.0
    # via -r requirements.in
click==8.3.1
    # via uvicorn
fastapi==0.123.0
    # via
    #   -r requirements.in
    #   strawberry-graphql
graphql-core==3.2.7
    # via strawberry-graphql
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
httptools==0.7.1
    # via uvicorn
idna==3.11
    # via anyio
lia-web==0.2.3
    # via strawberry-graphql
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
packaging==25.0
    # via strawberry-graphql
psycopg[binary]==3.2.13
    # via -r requirements.in
psycopg-binary==3.2.13
    # via psycopg
pydantic==2.12.5
    # via fastapi
pydantic-core==2.41.5
    # via pydantic
python-dateutil==2.9.0.post0
    # via strawberry-graphql
python-dotenv==1.2.1
    # via uvicorn
python-multipart==0.0.20
    # via strawberry-graphql
pyyaml==6.0.3
    # via uvicorn
redis==7.1.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.44
    # via
    #   -r requirements.in
    #   alembic
starlette==0.50.0
    # via fastapi
strawberry-graphql[fastapi]==0.287.0
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   lia-web
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   strawberry-graphql
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uvicorn[standard]==0.38.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
